% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fe.R
\name{lm.fe}
\alias{lm.fe}
\title{Linear regression with fixed-effects}
\usage{
lm.fe(formula = y ~ x, data, demean.data = NULL, wgt = NULL, fevar, ...)
}
\arguments{
\item{...}{}
}
\value{
A lm() object
}
\description{
Linear regression with fixed-effects
}
