% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lsa.R
\name{lm.lsa}
\alias{lm.lsa}
\title{Run regression analyses with ILSA data}
\usage{
lm.lsa(
  formula = y ~ x,
  data,
  wgt,
  fevar = NULL,
  rwgts,
  study,
  pvs,
  ncores = NULL,
  benchmark = FALSE
)
}
\arguments{
\item{formula}{model formula}

\item{data}{a data frame}

\item{wgt}{a string indicating the column in the data with the total weights}

\item{fevar}{a string indicating the column(s) in the data with the variables
used as fixed-effects. Supports 1 or more. If NULL, then there is no fixed-effect adjustment}

\item{rwgts}{a vector of strings indicating the names of the replicate weights variables
(e.g., if 150 replicates are wanted, then the vector must be length 150)}

\item{study}{a string indicating study, supports: 'TIMSS', 'PIRLS','ICILS',"ICCS', and 'PISA'}

\item{pvs}{a named list.
Each list element name indicates how is referred within the formula
Each list element must contain a vector of strings indicating the name in the dataset}

\item{ncores}{indicate desired number of cores used for processing.
1 for no parallel processing.
If NULL (default), the estimation will use half of the logical cores available}

\item{benchmark}{TRUE if want to estimate the time used per each plausible value estimation}
}
\value{
a summary.lm() object
}
\description{
Run regression analyses with replicate weights and combining plausible values.
Supports TIMSS, PIRLS, ICILS, ICCS, and PISA.
Does not support factor() and interactions yet
}
\examples{

# Example without fixed-effects
data.deu.2011 <- mini_pirls[mini_pirls$YEAR \%in\% "2011" &
                            mini_pirls$IDCNTRY \%in\% 276,]
example1 <- lm.lsa(ASRREA ~ 1 + PRESCH, # formula
                      data=data.deu.2011, # example data
                      wgt="SENWGT", # weight: senate weights
                      fevar=NULL, # no fixed-effects
                      rwgts=paste0("RWGT",1:150), # name of replicate weights
                      pvs=list(ASRREA=paste0("ASRREA0",1:5)), # list with PVs variables
                      ncores=NULL, # automatic ncores
                      study="PIRLS"
)
print(example1)


# Example with time and country fixed-effects
example2 <- lm.lsa(ASRREA ~ 1 + PRESCH, # formula
                      data=mini_pirls, # example data
                      wgt="SENWGT", # weight: senate weights
                      fevar=c("YEAR","IDCNTRY"), # fixed-effects variables
                      rwgts=paste0("RWGT",1:150), # name of replicate weights
                      pvs=list(ASRREA=paste0("ASRREA0",1:5)), # list with PVs variables
                      ncores=NULL, # automatic ncores
                      study="PIRLS"
)
print(example2)



}
