% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_rep.R
\name{lm.rep}
\alias{lm.rep}
\title{Title}
\usage{
lm.rep(
  formula = y ~ x,
  data,
  wgt,
  fevar = NULL,
  rwgts,
  study,
  ncores = NULL,
  benchmark = TRUE,
  asList = FALSE
)
}
\arguments{
\item{formula}{model formula}

\item{data}{a data frame}

\item{wgt}{a string indicating the column in the data with the total weights}

\item{fevar}{a string indicating the column(s) in the data with the variables
used as fixed-effects. Supports 1 or more. If NULL, then there is no fixed-effect adjustment}

\item{rwgts}{a vector of strings indicating the names of the replicate weights variables
(e.g., if 150 replicates are wanted, then the vector must be length 150)}

\item{study}{a string indicating study, supports: 'TIMSS', 'PIRLS','ICILS',"ICCS', and 'PISA'}

\item{ncores}{indicate desired number of cores used for processing.
1 for no parallel processing.
If NULL (default), the estimation will use half of the logical cores available}

\item{benchmark}{TRUE if want to estimate the time used per each plausible value estimation}

\item{asList}{if FALSE (default), returns summary() of the regression
if TRUE, return a list with some original lm() elements (such as the dataframe)}
}
\value{
a summary.lm() object
}
\description{
Title
}
\examples{

# Example without fixed-effects, dependent variable 1st plausible value
data.deu.2011 <- mini_pirls[mini_pirls$YEAR \%in\% "2011" &
                            mini_pirls$IDCNTRY \%in\% 276,]
example1 <- lm.rep(ASRREA01 ~ 1 + PRESCH, # formula
                      data=data.deu.2011, # example data
                      wgt="SENWGT", # weight: senate weights
                      fevar=NULL, # no fixed-effects
                      rwgts=paste0("RWGT",1:150), # name of replicate weights
                      ncores=NULL, # automatic ncores
                      study="PIRLS"
)
print(example1)


# Example with time and country fixed-effects, dependent variable 1st plausible value
example2 <- lm.lsa(ASRREA01 ~ 1 + PRESCH, # formula
                      data=mini_pirls, # example data
                      wgt="SENWGT", # weight: senate weights
                      fevar=c("YEAR","IDCNTRY"), # fixed-effects variables
                      rwgts=paste0("RWGT",1:150), # name of replicate weights
                      ncores=NULL, # automatic ncores
                      study="PIRLS"
)
print(example2)


}
